@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <Configurations>Debug</Configurations>
         <Platforms>x64</Platforms>
         <Nullable>enable</Nullable>
