@@ -9,6 +9,7 @@ using Cethleann.Structure.DataStructs;
 using DragonLib;
 using DragonLib.CLI;
 using DragonLib.IO;
+using System.Net.Http;
 
 namespace Cethleann.Downloader
 {
@@ -51,10 +52,17 @@ namespace Cethleann.Downloader
                 var (url, dest, size) = pair;
                 Logger.Info("Cethleann", url);
                 if (flags.Dry) return;
-                using var http = new WebClient();
+                using var http = new HttpClient();
                 try
                 {
-                    http.DownloadFile(url, dest);
+                    var folder = Path.GetDirectoryName(dest);
+                    if (!string.IsNullOrEmpty(folder) && !Directory.Exists(folder))
+                    {
+                        Directory.CreateDirectory(folder);
+                    }
+
+                    using var stream = http.GetStreamAsync(url);
+                    using var destStream = File.OpenWrite(dest);
                     Logger.Info("Cethleann", $"Downloaded {size} to {dest}");
                 }
                 catch (Exception e)
